Rename files to lowercase and update import statements; replace print with stdout.write for debug messages
